+++ app/src/main/java/ru/empat/morewords/domain/entity/Word.kt
@@ -1,8 +1,8 @@
 package ru.empat.morewords.domain.entity
 
 data class Word(
-    var wordId: Int,
-    var dictionaryId: Int,
+    var wordId: Long,
+    var dictionaryId: Long,
     var text: String,
     var translate: String
 )
